Fix React Hooks violation in tile rendering
Created ConnectedOverviewTile wrapper component to properly handle
calling tile data hooks. This fixes the "Rendered more hooks than
during the previous render" error.

Issue: Calling tile.useTileData() inside map function violated Rules
of Hooks because the number of tiles changes based on permissions.

Solution: Each tile now gets its own component instance that calls
the hook unconditionally, ensuring consistent hook ordering.
